extracted time_format() to its own file
